Create Loop exercise 1
When you run this program, it will iterate from 0 to 15 and for each number, it will check if it's even or odd using the modulo operator (%). If the number is divisible by 2 (i.e., the remainder is 0), it is even; otherwise, it is odd. The program then displays a message indicating whether each number is even or odd.