Create Loop Exercise 2
When you run this program, it will prompt the user to input a number. After the user provides the input, the program will iterate from 1 to 10 and calculate the result of multiplying the given number with each iteration value. It will then display each entry of the multiplication table using console.log().