Create Vanilla Array 2
In this program, we start with the given array [20, 30, 25, 35, -16, 60, -100]. We initialize a variable sum to 0. Then, we iterate through each element in the array using a for loop. In each iteration, we add the current element to the sum variable. After the loop, we calculate the average by dividing the sum by the number of elements in the array (array.length). Finally, we display the calculated average using console.log().