Create Vanilla function 1
In this program, we have two functions:

showArrayContent: This function takes an array as a parameter and displays the content of the array using console.log().

findMaxValue: This function takes an array as a parameter and iterates through the array to find the maximum value. It initializes the max variable with the first element of the array and compares each subsequent element with the current maximum value. If a larger value is found, it updates the max variable accordingly. Finally, it displays the maximum value using console.log().

We define the sample array [10, 2, 3, 4, 7] and call both functions, passing the array as an argument.